Move the objects to the correct location
