+++ types/inquirer/lib/objects/choice.d.ts
@@ -1,9 +1,9 @@
-import inquirer = require('..');
+import inquirer = require('../..');
 
 /**
  * Represents a choice for several question-types.
  */
-export class Choice<T extends inquirer.poll.Answers = inquirer.poll.Answers> {
+declare class Choice<T extends inquirer.poll.Answers = inquirer.poll.Answers> {
     /**
      * The name of the choice.
      */
@@ -35,3 +35,5 @@ export class Choice<T extends inquirer.poll.Answers = inquirer.poll.Answers> {
      */
     public constructor(value: any, answers: T);
 }
+
+export = Choice;
